requirements.txt
numpy~=1.18.1
matplotlib~=3.1.3
pytorch~=1.5.0
torchvision~=0.6.0
gym~=0.17.2
atari-py~=1.2.1
pillow~=7.0.0
ipython~=7.12.0
pyyaml~=5.3
yaml~=0.1.7